requirements.txt
flask==3.0.0
flask-restful==0.3.10
mysqlclient==2.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0